bot/requirements.txt
discord.py>=2.4.1
python-dotenv>=1.1.0
requests>=2.31.0
